% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAoutlierRemover-package.R
\docType{package}
\name{PCAoutlierRemover-package}
\alias{PCAoutlierRemover}
\alias{PCAoutlierRemover-package}
\title{PCAoutlierRemover: Interactive outlier removal from PCA}
\description{
Provides an interactive Shiny gadget for identifying and removing
outlier samples from 2D PCA scatter plots of population genomic data.
}
\section{Typical workflow}{

\enumerate{
\item Filter your VCF (e.g. with SNPfiltR)
\item Convert to genind with \code{vcfR::vcfR2genind()}
\item Call \code{compute_pca()} to get scores
\item Call \code{remove_outliers()} to launch the gadget
\item Apply the returned ID list to filter your genind/vcfR
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jhereward/PCAoutlierRemover}
  \item Report bugs at \url{https://github.com/jhereward/PCAoutlierRemover/issues}
}

}
\author{
\strong{Maintainer}: James Hereward \email{j.hereward@uq.edu.au}

Authors:
\itemize{
  \item James Hereward \email{j.hereward@uq.edu.au}
}

}
\keyword{internal}
